% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/m3plot.R
\name{m3plot}
\alias{m3plot}
\title{A function to plot the maximum, mean and minimum values of a numerical factor}
\usage{
m3plot(x)
}
\arguments{
\item{x}{is the input vector}
}
\value{
plot
}
\description{
A function to plot the maximum, mean and minimum values of a numerical factor
}
\examples{
m3plot(mtcars$hp)
}

